+++ src/App.js
@@ -1,13 +1,12 @@
 import './App.css';
-
+import HomePage from '../src/Pages/HomePage'
 
 function App() {
   return (
 <div className="main-div">
-
+    <HomePage/>
     </div>
   );
 }
 
 export default App;
-``
