+++ src/Pages/HomePage.js
@@ -1,13 +1,31 @@
 import React from 'react';
-import Home from '../components/Homepage/Home'
-
+import Hero from '../components/Homepage/Hero'
+import About from '../components/Homepage/About'
+import Cards from '../components/Homepage/Cards'
+import Viewbtn from '../components/Homepage/Viewbtn'
 
 const HomePage = () => {
   return (
     
     <>
     
-<Home/>
+    <div className="first-div">
+    <div className="second-div">
+
+
+    <Hero/>
+    <About/>
+    <Cards/>
+    <Viewbtn/>
+
+
+    </div>
+
+
+</div>
+
+
+
 
     </>
   );
