+++ src/components/Homepage/About.js
@@ -1,0 +1,32 @@
+import React from 'react'
+
+
+export default function About() {
+    return (
+        <>
+
+            <div className="about-div">
+                <div className="about-img">
+                    <img src="/Home/Aboutimg.png" alt="about-image" />
+                </div>
+
+                <div className="about-text-div">
+                    <h1 className="my-10 text-white">ABOUT</h1>
+                    <p className=" text-grey p-about">
+                        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam
+                    </p>
+                </div>
+
+
+
+            </div>
+
+
+
+
+
+
+
+        </>
+    )
+}
