+++ src/components/Homepage/Hero.js
@@ -1,0 +1,18 @@
+import React from 'react'
+export default function Hero() {
+    return (
+        <>
+            <div className="hero">
+                <p className="">Voices form the Cosmos:</p>
+                <p>Spaces Meets Imagination and</p>
+                <p>Discovery</p>
+
+                <div className="input-div gap-5 ">
+                    <input type="text" className="search-input" />
+                    <button className="search-button"><img src="/Home/vector.png" alt="icon" class="button-icon w-10 h-10" /></button>
+                </div>
+            </div>
+
+        </>
+    )
+}
