+++ src/components/Homepage/Viewbtn.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+export default function Viewbtn() {
+  return (
+    <>
+    
+    <div className="View-post-div">
+
+                    <button className="view-post-btn">View All Posts</button>
+
+                </div>
+    
+    
+    </>
+  )
+}
