+++ src/index.css
@@ -2,8 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
-
-
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
@@ -17,3 +15,188 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+/* Main background div */
+.main-div {
+  margin: 0;
+  padding: 0;
+  height: 100%;
+  width: 100%;
+  background: radial-gradient(circle at top left, 
+      rgba(211, 150, 105, 0.9), 
+      rgba(0, 0, 30, 0.9) 15%, 
+      rgba(0, 0, 0, 1) 30%),
+    linear-gradient(180deg, rgba(0, 15, 48, 0.4) 0%, rgba(0, 48, 153, 0.4) 90%);
+  backdrop-filter: blur(100px);
+  display: flex;
+  justify-content: center;
+}
+
+/* first div in home.js */
+.first-div {
+  margin-top: 20px;
+  width: 80%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  margin-bottom: 20px;
+}
+/*  p element css for hero section*/
+.second-div p {
+  text-align: center;
+  color: white;
+  font-family: "Orbitron", sans-serif;
+  font-weight: 800;
+  font-size: 34px;
+}
+
+/* Input div for the button and input field */
+.input-div {
+  margin-top: 50px;
+  width: 100%;
+  height: 65px;
+  justify-content: center;
+  display: flex;
+}
+
+/* Input filed search button */
+
+.search-button {
+  display: flex; 
+  align-items: center; 
+  justify-content: center; 
+  padding: 20px 30px; 
+  border: none; 
+  border-radius: 8px; 
+  background-color: rgba(217, 217, 217, 1); 
+  cursor: pointer;
+}
+
+/* search input css for input field*/
+
+.search-input {
+  background-color: rgba(217, 217, 217, 1);
+  border-radius: 10px;
+  width: 500px;
+}
+
+/* About div */
+.about-div {
+  margin-top: 200px;
+  width: 100%;
+  height: 527px;
+  display: inline-flex;
+  flex-direction: row;
+}
+
+.about-img {
+  width: 70%;
+  height: 100%;
+}
+
+.about-text-div {
+  width: 30%;
+  height: 100%;
+  font-family: 'plus jakarta sans', sans-serif;
+  color: grey;
+}
+
+.about-div p {
+  font-family: 'inter', sans-serif;
+  font-weight: 100;
+  font-size: 15px;
+  margin-top: 80px;
+}
+
+.cards-div {
+  margin-top: 50px;
+  width: 100%;
+  height: 960px;
+  
+}
+
+.cards-div p {
+  margin-bottom: 20px;
+  font-size: 50px;
+}
+
+.View-post-div {
+  width: 100%;
+  height: 50px;
+  margin-top: 20px;
+  margin-bottom: 50px;
+  display: flex;
+  justify-content: center;
+}
+
+.view-post-btn {
+  padding: 10px;
+  font-family: 'work-sans', sans-serif;
+  color: rgba(105, 106, 117, 1);
+  height: 100%;
+  border: 1px solid grey;
+  border-radius: 6px;
+}
+
+/* Responsive Design */
+
+/* For small devices (phones) */
+@media (max-width: 768px) {
+  .about-div {
+    flex-direction: column;
+    height: auto;
+  }
+
+  .about-img {
+    width: 100%;
+    height: auto;
+  }
+
+  .about-text-div {
+    width: 100%;
+    height: auto;
+    margin-top: 20px;
+    text-align: center;
+  }
+
+  .search-input {
+    width: 90%;
+  }
+
+  .cards-div {
+    height: auto;
+  }
+
+  .cards-div p {
+    font-size: 30px;
+  }
+}
+
+
+
+/* For medium devices (tablets) */
+@media (max-width: 1024px) {
+  .about-div {
+    flex-direction: column;
+    height: auto;
+  }
+
+  .about-img {
+    width: 100%;
+    height: auto;
+  }
+
+  .about-text-div {
+    width: 100%;
+    text-align: center;
+  }
+
+  .search-input {
+    width: 70%;
+  }
+
+  .cards-div {
+    height: auto;
+  }
+}
+
